Use always float as the floating point type for compositing (http://lists.kde.org/?l=kwin&m=118493073424327&w=2).
svn path=/trunk/KDE/kdebase/workspace/; revision=700026

+++ deleted.h
@@ -35,7 +35,7 @@ class Deleted
         void copyToDeleted( Toplevel* c );
         virtual ~Deleted(); // deleted only using unrefWindow()
         int delete_refcount;
-        double window_opacity;
+        float window_opacity;
         int desk;
         QRect contentsRect; // for clientPos()/clientSize()
     };
